renamed the config to client context to be consistent with other SDKs. Updated the docs.
