+++ sdks/typescript/client-typescript-core/app/client_feature_repository.ts
@@ -16,7 +16,7 @@ import {
   SSEResultState,
   FeatureStateTypeTransformer
 } from './models';
-import { ClientContext } from './config';
+import { ClientContext } from './client_context';
 
 export enum Readyness {
   NotReady = 'NotReady',
@@ -44,7 +44,7 @@ export class ClientFeatureRepository {
   // indexed by id
   private _catchReleaseStates = new Map<string, FeatureState>();
   private _newFeatureStateAvailableListeners: Array<PostLoadNewFeatureStateAvailableListener> = [];
-  private _config = new ClientContext();
+  private _clientContext = new ClientContext();
 
   public get readyness(): Readyness {
     return this.readynessState;
@@ -152,8 +152,8 @@ export class ClientFeatureRepository {
     return holder;
   }
 
-  get config(): ClientContext {
-    return this._config;
+  get clientContext(): ClientContext {
+    return this._clientContext;
   }
 
   get catchAndReleaseMode(): boolean {
